% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banxico.ts.R
\name{banxico.ts}
\alias{banxico.ts}
\title{Banxico Time Series Import}
\usage{
banxico.ts(series_id_banxico, initial_date, final_date)
}
\arguments{
\item{final_date}{}
}
\value{

}
\description{
Banxico Time Series Import
}
