% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Money_demand.R
\docType{data}
\name{Money_demand}
\alias{Money_demand}
\title{some money demand variables for Mexico}
\format{

}
\usage{
Money_demand
}
\description{
some money demand variables for Mexico
}
\keyword{datasets}
